# Fixed #33546 -- Optimized ResponseHeaders and HttpResponseBase