[ticket is 33546](https://code.djangoproject.com/ticket/33546#ticket).

Lets hope CI doesn't disprove me.

Timings, mirrored from ticket, on my local Macbook Air:

|                                                                	| main    	| patched 	|
|----------------------------------------------------------------	|---------	|---------	|
| ResponseHeaders({})                                            	| 745 ns  	| 313 ns  	|
| ResponseHeaders({'TEST': "1"})                                 	| 2.17 µs 	| 1.8 µs  	|
| ResponseHeaders({'TEST': "1", "TEST2": "2"})                   	| 3.51 µs 	| 2.77 µs 	|
| ResponseHeaders({"Content-Type": "application/json"})          	| 2.19 µs 	| 1.93 µs 	|
| HttpResponseBase()                                             	| 7.89 µs 	| 4.18 µs 	|
| HttpResponseBase(headers={"TEST": "1"})                        	| 9.75 µs 	| 6.02 µs 	|
| HttpResponseBase(headers={"TEST": "1", "TEST2": "2"})          	| 11 µs   	| 7.09 µs 	|
| HttpResponseBase(headers={"Content-Type": "application/json"}) 	| 3.71 µs 	| 3.26 µs 	|
| HttpResponseBase(content_type="test")                          	| 4.48 µs 	| 3.03 µs 	|
| HttpResponse()                                                 	| 9.06 µs 	| 5.46 µs 	|
| HttpResponse(headers={"TEST": "1"})                            	| 11.4 µs 	| 7.58 µs 	|
| HttpResponse(headers={"TEST": "1", "TEST2": "2"})              	| 13 µs   	| 8.83 µs 	|
| HttpResponse(headers={"Content-Type": "application/json"})     	| 5.22 µs 	| 4.82 µs 	|
| HttpResponse(content_type="test")                              	| 6.2 µs  	| 4.38 µs 	|

As always, I'd welcome anyone point out flaws,  additional timings to check, or counterpoint datapoints. Also scrutinise my logic, as I've been staring at this for _too long_ ;)